+++ packages/patrol_cli/lib/src/commands/build_ios.dart
@@ -82,7 +82,7 @@ class BuildIOSCommand extends PatrolCommand {
 
     // Check compatibility between CLI and package versions
     final patrolVersion = _pubspecReader.getPatrolVersion();
-    await _compatibilityChecker.checkVersionsCompatibilityWithWarning(
+    await _compatibilityChecker.checkVersionsCompatibilityForBuild(
       patrolVersion: patrolVersion,
     );