+++ packages/patrol_cli/lib/src/commands/build_macos.dart
@@ -78,7 +78,7 @@ class BuildMacOSCommand extends PatrolCommand {
 
     // Check compatibility between CLI and package versions
     final patrolVersion = _pubspecReader.getPatrolVersion();
-    await _compatibilityChecker.checkVersionsCompatibilityWithWarning(
+    await _compatibilityChecker.checkVersionsCompatibilityForBuild(
       patrolVersion: patrolVersion,
     );